v1.0.3b(7)
- search functionality, regression issue fixed
- comment sorting + new picker selector
- updated randomizer for instances
- feed should reset upon new instance connection
- config settings regression issue fixed
- removed auto correction for add to profiles
